backend/requirements.txt
fastapi
uvicorn
pydantic[email]
python-dotenv

# LangChain + Gemini
langchain
langchain-google-genai
google-generativeai
